b/f core: fix for CPS limiter fix for commit a7383254aba89059a4b22be48b4d471dec67f0f1, rounding error

+++ core/AmSessionContainer.cpp
@@ -404,7 +404,7 @@ void AmSessionContainer::setCPSSoftLimit(unsigned int percent)
       break;
     }
   }
-  CPSLimit = (percent / 100) * ((float)cps_queue.size() / CPS_SAMPLERATE);
+  CPSLimit = ((float)percent / 100) * ((float)cps_queue.size() / CPS_SAMPLERATE);
   if(0 == CPSLimit) CPSLimit = 1;
 }
